ai-dapseq/requirements.txt
biopython
pyBigWig
pandas
numpy
scipy
torch
scikit-learn
matplotlib
seaborn